content/canvas/test/test_2d.line.invalid.miterLimit.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.line.invalid.miterLimit</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ctx.miterLimit = 1.5;
ok(ctx.miterLimit === 1.5, "ctx.miterLimit === 1.5");

ctx.miterLimit = 1.5;
ctx.miterLimit = 0;
todo(ctx.miterLimit === 1.5, "ctx.miterLimit === 1.5");

ctx.miterLimit = 1.5;
ctx.miterLimit = -1;
todo(ctx.miterLimit === 1.5, "ctx.miterLimit === 1.5");

ctx.miterLimit = 1.5;
var _thrown = undefined; try {
  ctx.miterLimit = Infinity;
} catch (e) { _thrown = e }; todo(_thrown && _thrown.code == DOMException.NOT_SUPPORTED_ERR, "should throw NOT_SUPPORTED_ERR");
ok(ctx.miterLimit === 1.5, "ctx.miterLimit === 1.5");

ctx.miterLimit = 1.5;
var _thrown = undefined; try {
  ctx.miterLimit = -Infinity;
} catch (e) { _thrown = e }; todo(_thrown && _thrown.code == DOMException.NOT_SUPPORTED_ERR, "should throw NOT_SUPPORTED_ERR");
ok(ctx.miterLimit === 1.5, "ctx.miterLimit === 1.5");

ctx.miterLimit = 1.5;
var _thrown = undefined; try {
  ctx.miterLimit = NaN;
} catch (e) { _thrown = e }; todo(_thrown && _thrown.code == DOMException.NOT_SUPPORTED_ERR, "should throw NOT_SUPPORTED_ERR");
ok(ctx.miterLimit === 1.5, "ctx.miterLimit === 1.5");

if (!_deferred) SimpleTest.finish();
});
</script>
